refactor(sdk): split runtime prompts into base, memory, and templates modules
