+++ packages/sdk/src/runtime/prompts/index.ts
@@ -1,0 +1,3 @@
+export { BASE_INSTRUCTIONS } from "./base";
+export { MEMORY_SECTION } from "./memory";
+export { TEMPLATES, renderSystemPrompt, type RenderSystemPromptOptions } from "./templates";
